Split shell raw shell tests
